FAT: minimize 64-bit arithmetic
Represent FAT file sizes and on-volume sector indices as 32-bit values after validating inputs at the VFS boundary.

Cache power-of-two sector and cluster shifts and masks, and use them throughout file I/O, directory iteration, cluster traversal, and mount calculations to eliminate divisions and multiplications.

Compute FAT entry offsets and iterator positions in 32 bits, retaining 64-bit arithmetic only for device byte offsets.

+++ Kernel/FSDrivers/FAT/FATClusterSectorIterator.h
@@ -30,7 +30,7 @@ struct FATClusterSectorIterator
     FATClusterSectorIterator(Ptr<FATVolume> volume, uint32_t cluster, uint32_t sector);
     void Set(uint32_t cluster, uint32_t sector);
     
-    off64_t         GetBlockSector();
+    uint32_t        GetBlockSector();
     size_t          GetRemainingContiguousSectorCount();
     KCacheBlockDesc GetBlock_(bool doLoad, size_t readAheadBlockCount = 1);
